+++ afew/filters/BaseFilter.py
@@ -19,50 +19,8 @@ from __future__ import print_function, absolute_import, unicode_literals
 
 import collections
 import logging
-import pkg_resources
 
 import notmuch
-
-
-RAISEIT = object()
-
-
-class FilterRegistry(object):
-    """
-    The FilterRegistry is responsible for returning
-    filters by key.
-    Filters get registered via entry points.
-    To avoid any circular dependencies, the registry loads
-    the Filters lazily
-    """
-    def __init__(self, filters):
-        self._filteriterator = filters
-
-    @property
-    def filter(self):
-        if not hasattr(self, '_filter'):
-            self._filter = {}
-            for f in self._filteriterator:
-                self._filter[f.name] = f.load()
-        return self._filter
-
-    def get(self, key, default=RAISEIT):
-        if default == RAISEIT:
-            return self.filter[key]
-        else:
-            return self.filter.get(key, default)
-
-    def __getitem__(self, key):
-        return self.get(key)
-
-    def keys(self):
-        return self.filter.keys()
-
-    def values(self):
-        return self.filter.values()
-
-    def items(self):
-        return self.filter.items()
 
 
 class Filter(object):
@@ -158,5 +116,3 @@ class Filter(object):
                         message.remove_tag(tag)
 
         self.flush_changes()
-
-all_filters = FilterRegistry(pkg_resources.iter_entry_points('afew.filter'))
