+++ afew/filters/HeaderMatchingFilter.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 from __future__ import print_function, absolute_import, unicode_literals
-from ..Filter import Filter
+from .BaseFilter import Filter
 
 import re
